made a custom hook!
